# Optimize performance and memory usage in dynamics computations