--- a/tests/test_api_frame.py
+++ b/tests/test_api_frame.py
@@ -30,6 +30,7 @@ def test_frame_index(jaxsim_models_types: js.model.JaxSimModel):
             < model.number_of_links()
         )
 
+    # See discussion in https://github.com/ami-iit/jaxsim/pull/280
     assert js.frame.names_to_idxs(
         model=model, frame_names=model.frame_names()
     ) == pytest.approx(jnp.arange(n_l, n_l + n_f))