--- a/tests/test_api_joint.py
+++ b/tests/test_api_joint.py
@@ -19,6 +19,7 @@ def test_joint_index(
         assert js.joint.name_to_idx(model=model, joint_name=joint_name) == idx
         assert js.joint.idx_to_name(model=model, joint_index=idx) == joint_name
 
+    # See discussion in https://github.com/ami-iit/jaxsim/pull/280
     assert js.joint.names_to_idxs(
         model=model, joint_names=model.joint_names()
     ) == pytest.approx(jnp.arange(model.number_of_joints()))