Enhance efficiency by normalizing quaternions, reducing memory footprint, and speeding up various dynamics computations. Replace `jnp.float32` with `float` for consistency. Avoid unnecessary calls to `.names_to_idxs` where possible.
> [!Note]
> See profiler traces at https://gist.github.com/flferretti/a20c7cd171c25dd784eef361246bd8cd

> [!Tip]
> - `js.model.generalized_free_floating_jacobian_derivative` used `jax.vmap` over `js.link.jacobian_derivative`, but a lot of quantities were unnecessarily computed multiple times. For this reason, I moved the major computation to `js.model` and left only the link extraction part in `js.link`. Some adjustments regarding the output velocity representation were necessary in order to avoid to call `js.link.velocity` inside the `vmap`ped function `_compute_row`

| Function                                                         | New Compile Time | Original Compile Time | New Run Time | Original Run Time |
|------------------------------------------------------------------|------------------|-----------------------|--------------|-------------------|
| `js.model.generalized_free_floating_jacobian_derivative`         | 643 ms           | 748 ms                | **479 μs**      | 6.07 ms           |
| `js.model.free_floating_mass_matrix`                             | 423 ms           | 456 ms                | **470 μs**       | 502 μs            |
| `js.model.forward_kinematics`                                    | 246 ms           | 255 ms                | **406 μs**       | 445 μs            |
| `js.model.generalized_free_floating_jacobian`                    | 311 ms           | 368 ms                | **455 μs**       | 473 μs            |
| `js.references.JaxSimModelReferences.build`                      | 98.3 ms          | 139 ms                | **558 μs**       | 498 μs            |
| `js.contact.collidable_point_kinematics`                         | 326 ms           | 332 ms                | **420 μs**       | 432 μs            |
| `data.base_orientation`                         | 14.3 ms          | 19.1 ms               | **10.2 μs**      | 14.6 μs           |
| `model.kin_dyn_parameters.joint_transforms_and_motion_subspaces` | 236 ms           | 241 ms                | **53.3 μs**      | 60.1 μs           |
| `references.link_forces`                                         | 12.7 ms          | 38.2 ms               | **20.9 μs**      | 22.1 μs           |

<!-- readthedocs-preview jaxsim start -->
----
📚 Documentation preview 📚: https://jaxsim--280.org.readthedocs.build//280/

<!-- readthedocs-preview jaxsim end -->