@diegoferigo 2024-11-06
> > Leaving a preliminary comment before going more in detail with a review.
> > Do you have any benchmark supporting the need to switch from names to indices? Using indices makes the code much more fragile and, if the indexing logic changes, there will be countless of places to update. The name-to-index helpers provide a centralized location to store this logic.
> > A very descriptive example of what I just said is the joint name to index map. There is a tricky +- 1 index, that at first sight this PR lacks. I'm not that supportive of gaining minor performance benefits if the price to pay is less maintainable and more error prone code. However, I'll be waiting for some benchmark, we can assess the situation more clearly with data.
> 
> I haven't removed the usage of `*_names`, but simply avoided to call `js.*.names_to_idxs` when `*_names` is not passed to the call. In that case, we explicitly need all of the `*` (joint/link/frame), so we can just use `model.number_of_*`

Yes you introduced the knowledge that you can use `arange` in that case. But as said, this change spreads around this piece of knowledge that might change in the future (what if we start supporting free-floating joints to unify the base-related section of RBDAs and the corresponding base joint get a zero index?). If there are no relevant performance consideration in place, why making the code more hard to read and error prone for future refactorings? (this is also a general comment, not necessarily related to just this change).