@flferretti 2024-11-07
> But as said, this change spreads around this piece of knowledge that might change in the future (what if we start supporting free-floating joints to unify the base-related section of RBDAs and the corresponding base joint get a zero index?).

I believe that the most intuitive behavior for the final user is that 
```python
js.link.names_to_indexes(joint_names=model.link_names(), **kwargs)
```
and 
```python
jnp.arange(model.number_of_links())
```
 should return the same result. If there will be any need to change this logic in the future, we will propagate the change over the whole library