@diegoferigo 2024-11-07
> > But as said, this change spreads around this piece of knowledge that might change in the future (what if we start supporting free-floating joints to unify the base-related section of RBDAs and the corresponding base joint get a zero index?).
> 
> I believe that the most intuitive behavior for the final user is that
> 
> ```python
> js.link.names_to_indexes(joint_names=model.link_names(), **kwargs)
> ```
> 
> and
> 
> ```python
> jnp.arange(model.number_of_links())
> ```
> 
> should return the same result. If there will be any need to change this logic in the future, we will propagate the change over the whole library

My point is not on the user side since users are not supposed to know anything about indices (there are the names to index helpers for this reason). My point was about the developers and the logic within JaxSim functions.

I'm not strongly against this change, my whole point is that in case of changes on the indices used internally, there will be the need to change multiple locations. I checked the tests, and we already have checks that ensure that the current indexing does not get broken:

https://github.com/ami-iit/jaxsim/blob/bd590609ab72f639e733446270e2b26a974904a6/tests/test_api_joint.py#L22-L24

https://github.com/ami-iit/jaxsim/blob/main/tests/test_api_link.py#L27-L29

Can you please add a comment to both files that links to our discussion in this PR in case that check starts to fail?