@flferretti 2024-11-07
> Can you please add a comment to both files that links to our discussion in this PR in case that check starts to fail?

Done in https://github.com/ami-iit/jaxsim/pull/280/commits/85ab07ca445c955c1f4c3dcdaa7d6ad75c396922. Thanks!