@diegoferigo 2024-11-08
> Traces available at https://gist.github.com/flferretti/a20c7cd171c25dd784eef361246bd8cd.
> 
> In order to view them, please run:
> 
> ```shell
> git clone https://gist.github.com/a20c7cd171c25dd784eef361246bd8cd.git
> cd a20c7cd171c25dd784eef361246bd8cd
> python -m http.server
> ```
> 
> C.C. @xela-95 @diegoferigo

Can you please summarize them, if you have any specific improvement in mind?