@flferretti 2024-11-08
> Can you please summarize them, if you have any specific improvement in mind?

I made the gist just for reference and in case you were interested into the specific improvements. For a general benchmark, you can take a look at the PR description